requirements.txt
Flask>=3.0
Pillow>=10.0
scikit-learn>=1.3
numpy>=1.24
